native expressions
Technically, a native function can be like a native expression, but there are basically two distinct features I need.
(1) a native expression that is opaque and requires the annotator to treat it like an unknown. This is for doing arbitrary computation.
(2) a native expression that's "pure" - like a Math.sin() function, where I want even the annotator to be able to use the expression.